pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phasegpt"
version = "1.4.0"
description = "PhaseGPT: Volitional AI Framework"
authors = [
    {name = "TempleTwo", email = "dev@templetwo.ai"},
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "peft>=0.5.0",
    "pydantic>=2.0.0",
    "pyyaml",
    "bitsandbytes",
    "accelerate",
    "safetensors"
]
requires-python = ">=3.9"

[tool.setuptools.packages.find]
where = ["src"]
